refactor(editor): Remove secret provider mock API usage and mock file
